Converted links in CONTRIBUTING.md to relative
Addressing issue #26, converted links in CONTRIBUTING.md to realtive links for the purpose of easier management of forks. Some of these might be re-evaluated later as we decide whether to scope this document for Caravan contribution at the master level, or at the local fork level. Also, changed some copy that still had the Transit Network name.